Switched folders for components and assets
