+++ code/src/assets/images/home-page/intro-banner/computer-image.jpg
