+++ code/src/assets/images/home-page/intro-banner/mobile-image.jpg
