+++ code/src/assets/images/home-page/intro-banner/tablet-image.jpg
