+++ code/src/layouts/categorySection/CategorySection.constants.ts
