+++ code/src/layouts/categorySection/CategorySection.scss
