+++ code/src/layouts/categorySection/messages/en.json
@@ -1,0 +1,6 @@
+{
+  "categorySection.title": "Shop By Category",
+  "categorySection.computers": "Computers",
+  "categorySection.mobile": "Mobile",
+  "categorySection.tablets": "Tablets"
+}
