+++ code/src/layouts/categorySection/messages/index.ts
@@ -1,0 +1,4 @@
+import en from "@/layouts/categorySection/messages/en.json";
+import uk from "@/layouts/categorySection/messages/uk.json";
+
+export default { en, uk };
