+++ code/src/layouts/categorySection/messages/uk.json
@@ -1,0 +1,6 @@
+{
+  "categorySection.title": "Купуйте за категорією",
+  "categorySection.computers": "Комп'ютери",
+  "categorySection.mobile": "Мобільний",
+  "categorySection.tablets": "Планшети"
+}
